* awn-manager/awnApplet.py * awn-manager/awnPreferences.py * awn-manager/window.glade
You can now install python applets via drag and drop on to the right applet pane.
The basic package should be:
	yourapplet.tar.gz
	  /yourapplet
	  /yourapplet.desktop
	  
You can delete applets that are installed in your home directory
Applet devs can now use a relative path in the .desktop file for the Icon as it will be replaced when the correct absolute path when installed.
A fix for awn crashing when an awn manager couldn't find the icon set in the .desktop file

Added the ablity to toggle name_change_notify